requirements.txt
Django>=2.1,<2.2
django-cors-headers==2.4.0
djangorestframework>=3.8,<3.9
pytz==2018.5
gunicorn==19.9.0